+++ packages/wallet-management/src/walletIcons/index.ts
@@ -1,23 +1,23 @@
 import alphawallet from './alphawallet.svg';
-import atoken from './atoken.svg';
 import binance from './binance.svg';
 import bitpie from './bitpie.svg';
 import blockwallet from './blockwallet.svg';
 import brave from './brave.svg';
 import coinbase from './coinbase.svg';
 import dcent from './dcent.svg';
+import exodus from './exodus.svg';
 import frame from './frame.svg';
-import huobiwallet from './huobiwallet.svg';
+import frontier from './frontier.svg';
 import hyperpay from './hyperpay.svg';
 import imtoken from './imtoken.svg';
 import liquality from './liquality.svg';
 import mathwallet from './mathWallet.svg';
-import meetone from './meetone.svg';
 import metamask from './metaMask.svg';
-import mykey from './mykey.svg';
 import oneInch from './oneInch.svg';
 import opera from './opera.svg';
 import ownbit from './ownbit.svg';
+import placeholder from './placeholder.svg';
+import safe from './safe.svg';
 import status from './status.svg';
 import tallyho from './tallyho.svg';
 import tokenary from './tokenary.svg';
@@ -26,18 +26,15 @@ import trust from './trust.svg';
 import walletConnect from './walletConnect.svg';
 import walletio from './walletio.svg';
 import xdefi from './xdefi.svg';
-import frontier from './frontier.svg';
-import placeholder from './placeholder.svg';
-import exodus from './exodus.svg';
-import safe from './safe.svg';
+import bitkeep from './bitkeep.svg';
 
 export const walletIcons = {
   mathwallet,
   walletConnect,
   alphawallet,
-  atoken,
   frontier,
   blockwallet,
+  bitkeep,
   binance,
   placeholder,
   bitpie,
@@ -45,13 +42,10 @@ export const walletIcons = {
   coinbase,
   dcent,
   frame,
-  huobiwallet,
   hyperpay,
   imtoken,
   liquality,
-  meetone,
   metamask,
-  mykey,
   oneInch,
   opera,
   ownbit,