+++ packages/wallet-management/src/wallets.ts
@@ -49,6 +49,16 @@ const walletConnect: Wallet = new WalletConnectConnector({
   },
 });
 
+const bitKeep: Wallet = new InjectedConnector(
+  {
+    name: 'BitKeep',
+    installed: async () =>
+      (window as any).bitkeep && (window as any).bitkeep?.ethereum,
+    icon: walletIcons.bitkeep,
+  },
+  (window as any).bitkeep?.ethereum,
+);
+
 const frontier: Wallet = new InjectedConnector(
   {
     name: 'Frontier',
@@ -132,13 +142,6 @@ const alphawallet: Wallet = new InjectedConnector({
   icon: walletIcons.alphawallet,
 });
 
-const atoken: Wallet = new InjectedConnector({
-  name: 'AToken',
-  installed: async () =>
-    (window as any).ethereum?.[ProviderIdentityFlag.AToken],
-  icon: walletIcons.atoken,
-});
-
 const apex: Wallet = new InjectedConnector({
   name: 'Apex Wallet',
   installed: async () =>
@@ -167,13 +170,6 @@ const frame: Wallet = new InjectedConnector(
   (window as any).frame,
 );
 
-const huobiwallet: Wallet = new InjectedConnector({
-  name: 'HuobiWallet',
-  installed: async () =>
-    (window as any).ethereum?.[ProviderIdentityFlag.HuobiWallet],
-  icon: walletIcons.huobiwallet,
-});
-
 const hyperpay: Wallet = new InjectedConnector({
   name: 'HyperPay',
   // Note: The property `hiWallet` is as of now the only known way of identifying hyperpay
@@ -198,19 +194,6 @@ const liquality: Wallet = new InjectedConnector(
   },
   (window as any).liquality,
 );
-
-const meetone: Wallet = new InjectedConnector({
-  name: 'MeetOne',
-  installed: async () =>
-    (window as any).ethereum?.[ProviderIdentityFlag.MeetOne] === 'MEETONE',
-  icon: walletIcons.meetone,
-});
-
-const mykey: Wallet = new InjectedConnector({
-  name: 'MyKey',
-  installed: async () => (window as any).ethereum?.[ProviderIdentityFlag.MyKey],
-  icon: walletIcons.mykey,
-});
 
 const ownbit: Wallet = new InjectedConnector({
   name: 'OwnBit',
@@ -288,6 +271,7 @@ export const supportedWallets = [
   safe,
   metamask,
   walletConnect,
+  bitKeep,
   exodus,
   tallyho,
   binance,
@@ -296,19 +280,15 @@ export const supportedWallets = [
   trust,
   status,
   alphawallet,
-  atoken,
   blockWallet,
   bitpie,
   brave,
   apex,
   dcent,
   frame,
-  huobiwallet,
   hyperpay,
   imtoken,
   liquality,
-  meetone,
-  mykey,
   ownbit,
   tokenpocket,
   xdefi,