+++ packages/widget/src/components/SendToWallet/SendToWalletButton.tsx
@@ -2,13 +2,15 @@ import WalletIcon from '@mui/icons-material/Wallet';
 import { Button, Tooltip } from '@mui/material';
 import { useFormContext } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
+import { useWidgetEvents } from '../../hooks';
 import { FormKey, useWallet, useWidgetConfig } from '../../providers';
 import { useSendToWalletStore, useSettings } from '../../stores';
-import { DisabledUI, HiddenUI, RequiredUI } from '../../types';
+import { DisabledUI, HiddenUI, RequiredUI, WidgetEvent } from '../../types';
 
 export const SendToWalletButton: React.FC = () => {
   const { t } = useTranslation();
   const { setValue } = useFormContext();
+  const emitter = useWidgetEvents();
   const { account } = useWallet();
   const { disabledUI, hiddenUI, requiredUI } = useWidgetConfig();
   const { showSendToWallet, toggleSendToWallet } = useSendToWalletStore();
@@ -28,6 +30,10 @@ export const SendToWalletButton: React.FC = () => {
       setValue(FormKey.ToAddress, '', { shouldTouch: true });
     }
     toggleSendToWallet();
+    emitter.emit(
+      WidgetEvent.SendToWalletToggled,
+      useSendToWalletStore.getState().showSendToWallet,
+    );
   };
 
   return (