+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [2.3.0](https://github.com/lifinance/widget/compare/v2.2.8...v2.3.0) (2023-08-25)
+
+
+### Features
+
+* add SendToWalletToggled event ([e6bffcf](https://github.com/lifinance/widget/commit/e6bffcf013b34aca999d6a2763a4866c4164b52b))
+
+### [2.2.8](https://github.com/lifinance/widget/compare/v2.2.7...v2.2.8) (2023-08-18)
+
+
+### Bug Fixes
+
+* remove Huobi, MeetOne, AToken and MyKey wallets (LF-4489) ([#120](https://github.com/lifinance/widget/issues/120)) ([3f52729](https://github.com/lifinance/widget/commit/3f52729a02b6824777707ddc13c65403cc179c09))
+
+### [2.2.7](https://github.com/lifinance/widget/compare/v2.2.6...v2.2.7) (2023-08-14)
+
+
+### Bug Fixes
+
+* chain order for multiple widget instances ([4255ea3](https://github.com/lifinance/widget/commit/4255ea3ef4c027b2c1a3fedd0004eca6b1db2a2b))
+
 ### [2.2.6](https://github.com/lifinance/widget/compare/v2.2.5...v2.2.6) (2023-08-02)